refactor: rename spreadsheet to xlsx
